+++ applications/notification-supplier/src/test/java/org/opendaylight/openflowplugin/applications/notification/supplier/impl/item/stat/FlowTableStatNotificationSupplierImplTest.java
@@ -37,12 +37,14 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.Nodes;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.nodes.Node;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.nodes.NodeKey;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 public class FlowTableStatNotificationSupplierImplTest {
 
     private static final String FLOW_NODE_ID = "openflow:111";
-    private static final Short FLOW_TABLE_ID = 111;
+    private static final Uint8 FLOW_TABLE_ID = Uint8.valueOf(111);
     private static final String FLOW_ID = "test-flow-111";
+
     private FlowTableStatNotificationSupplierImpl notifSupplierImpl;
     private NotificationPublishService notifProviderService;
     private DataBroker dataBroker;
